src/requirements.txt
django
djangorestframework
drf-yasg
openapi
psycopg2-binary
celery